+++ apps/Standalone/src/desginer/app/AzureLogicAppsDesigner/laDesigner.tsx
@@ -58,6 +58,7 @@ import isEqual from 'lodash.isequal';
 import { useEffect, useMemo, useState } from 'react';
 import type { QueryClient } from 'react-query';
 import { useDispatch, useSelector } from 'react-redux';
+import { Chatbot } from '@microsoft/logic-apps-chatbot';
 
 const apiVersion = '2020-06-01';
 const httpClient = new HttpClient();
@@ -179,7 +180,6 @@ const DesignerEditor = () => {
   }, [data?.properties.files]);
 
   if (isLoading || appLoading || settingsLoading || customCodeLoading) {
-    // eslint-disable-next-line react/jsx-no-useless-fragment
     return <></>;
   }
 
@@ -328,7 +328,7 @@ const DesignerEditor = () => {
                 }}
               />
               <Designer rightShift={showChatBot ? 'chatbotPanelWidth' : undefined} />
-              {/* {showChatBot ? (
+              {showChatBot ? (
                 <Chatbot
                   openAzureCopilotPanel={() => openPanel('Azure Copilot Panel has been opened')}
                   getAuthToken={getAuthToken}
@@ -336,7 +336,7 @@ const DesignerEditor = () => {
                   openFeedbackPanel={() => openPanel('Azure Feedback Panel has been opened')}
                   closeChatBot={() => dispatch(setIsChatBotEnabled(false))}
                 />
-              ) : null} */}
+              ) : null}
             </div>
           </BJSWorkflowProvider>
         ) : null}
