+++ apps/Standalone/tsconfig.json
@@ -5,7 +5,8 @@
     "lib": [
       "ES2020",
       "DOM",
-      "DOM.Iterable"
+      "DOM.Iterable",
+      "ES2021.String"
     ],
     "module": "ESNext",
     "skipLibCheck": true,
